requirement.txt
clip==1.0
matplotlib==3.8.0
numpy==1.26.0
pandas==2.1.1
Pillow==10.0.1
PyYAML 
scikit-image==0.21.0
scikit-learn==1.3.1
scipy==1.11.2
tensorboard==2.15.1
tensorboardX==2.6.2.2
timm==0.6.13
tokenizers==0.14.1
torch==2.0.1
torchsummary==1.5.1
torchvision==0.15.2
tqdm==4.66.1
transformers==4.35.0
wandb==0.15.12

